feat(04-03): add metrics endpoint with token auth and unit tests
- Add METRICS_API_TOKEN optional field to env-validation.ts (Zod schema)
- Add production security warning for short METRICS_API_TOKEN
- Create /api/security/metrics endpoint with crypto.timingSafeEqual auth
- Returns 403 when token not configured, missing, wrong, or different length
- Returns metrics JSON with Cache-Control: no-cache headers on success
- 7 unit tests covering all auth scenarios (140 tests total)
