+++ src/lib/env-validation.ts
@@ -37,6 +37,10 @@ const envSchema = z.object({
   ADMIN_API_TOKEN: z.string()
     .min(32, 'ADMIN_API_TOKEN must be at least 32 characters')
     .optional(),
+  // Metrics API authentication
+  METRICS_API_TOKEN: z.string()
+    .min(32, 'METRICS_API_TOKEN must be at least 32 characters')
+    .optional(),
   // Site URL validation for CSP and CORS
   NEXT_PUBLIC_SITE_URL: z
     .url('NEXT_PUBLIC_SITE_URL must be a valid URL')
@@ -102,7 +106,11 @@ export function performSecurityChecks(env: EnvConfig): void {
     if (env.ADMIN_API_TOKEN && env.ADMIN_API_TOKEN.length < 64) {
       warnings.push('Production ADMIN_API_TOKEN should be at least 64 characters')
     }
-    
+
+    if (env.METRICS_API_TOKEN && env.METRICS_API_TOKEN.length < 64) {
+      warnings.push('Production METRICS_API_TOKEN should be at least 64 characters')
+    }
+
     if (!env.NEXT_PUBLIC_SITE_URL?.startsWith('https://')) {
       errors.push('Production site URL must use HTTPS')
     }
